# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Para servir y administrar restaurantes\nhttps://devsstore.net\n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.2.0",
    "title": "MenuXD",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://104.248.75.235:5000",
  "basePath": "/api/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/tables/": {
      "post": {
        "description": "Crea un nueva mesa",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Creado"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/tables/client/{clientId}": {
      "get": {
        "description": "Retorna un arreglo de mesas por su id",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "ID para único del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Table"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/tables/{id}": {
      "get": {
        "description": "Retorna una mesa por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener una mesa.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "description": "Actualiza una mesa por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para actualizar una mesa.",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "description": "Elimina una mesa por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para borrar una mesa.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/promotions/": {
      "post": {
        "description": "Crea una nueva promoción",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Promotion"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Creado"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/promotions/client/{clientId}": {
      "get": {
        "description": "Retorna un arreglo de promociones por su id.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "ID para único del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Promotion"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/promotion/{id}": {
      "get": {
        "description": "Retorna una promoción por su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para obtener una promoción.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Promotion"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "description": "Actualiza una promoción por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para actualizar una promoción.",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Promotion"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "description": "Elimina una promoción por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para borrar una promoción.",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Table": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2.0
        },
        "number": {
          "type": "number",
          "example": 10.0
        },
        "available": {
          "type": "boolean",
          "example": true
        },
        "type": {
          "type": "string",
          "example": "bar"
        },
        "client_id": {
          "type": "number",
          "example": 1.0
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
    "Promotion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2.0
        },
        "title": {
          "type": "string",
          "example": "2 x 1"
        },
        "description": {
          "type": "string"
        },
        "dish_id": {
          "type": "number",
          "example": "true"
        },
        "client_id": {
          "type": "number",
          "example": 2.0
        },
        "start_at": {
          "type": "string"
        },
        "end_at": {
          "type": "string"
        },
        "days": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
  }
}